update: move colors const to constants folder
